feat: adicionado logs em TODAS as ações do sistema web, os logs podem ser verificados, filtrados e exportados para .csv atrás do painel do admin. Os logs possuem a nossa função de paginação que será implementada no restante do sistema nos próximos commits

+++ v3.0/autoagenda_v3/src/main/java/br/com/autoagenda/autoagenda/controller/web/C_Veiculo.java
@@ -10,16 +10,17 @@ import br.com.autoagenda.autoagenda.model.Oficina;
 import br.com.autoagenda.autoagenda.model.Veiculo;
 import br.com.autoagenda.autoagenda.repositorios.ClienteRepository;
 import br.com.autoagenda.autoagenda.repositorios.VeiculoRepository;
+import br.com.autoagenda.autoagenda.service.LogService;
 
 @Controller
 @RequestMapping("/{slug}/veiculo-api")
 public class C_Veiculo {
+	@Autowired private LogService log;
     @Autowired private VeiculoRepository repo;
     @Autowired private ClienteRepository repoCliente;
     
     @GetMapping("/listar-por-cliente/{idCliente}") @ResponseBody
-    public ResponseEntity<List<Veiculo>> listarPorCliente(@SessionAttribute("oficinaAtual") Oficina oficina,
-    														@PathVariable Integer idCliente) {
+    public ResponseEntity<List<Veiculo>> listarPorCliente(@SessionAttribute("oficinaAtual") Oficina oficina, @PathVariable Integer idCliente) {
         List<Veiculo> veiculos = repo.findByCliente_OficinaAndAtivoTrue(oficina);
         return ResponseEntity.ok(veiculos);
     }
@@ -30,6 +31,7 @@ public class C_Veiculo {
             return ResponseEntity.badRequest().build();
         }
         
+        boolean novo = veiculo.getIdVeiculo() == null;
         Cliente cl = repoCliente.findById(veiculo.getCliente().getIdCliente()).orElseThrow();
         Veiculo veiculoSalvo;
         
@@ -49,15 +51,17 @@ public class C_Veiculo {
             veiculo.setCliente(cl);
             veiculoSalvo = repo.save(veiculo);
         }
+        log.registrar(novo ? "Criação" : "Edição", "Veículo", veiculoSalvo.getIdVeiculo(),
+                "Placa: " + veiculoSalvo.getPlaca()+ " | Modelo: " + veiculoSalvo.getModelo(), false);
         return ResponseEntity.ok(veiculoSalvo);
     }
     
     @PostMapping("/apagar") @ResponseBody
     public ResponseEntity<?> apagar(@RequestParam Integer idVeiculo) {
         try {
-            if(idVeiculo != null) { 
-                repo.deleteById(idVeiculo); 
-            }
+            if(idVeiculo != null) repo.deleteById(idVeiculo);
+            
+            log.registrar("Exclusão", "Veículo",idVeiculo, "Veículo removido ID: " + idVeiculo, false);
             return ResponseEntity.ok().build();
         } catch (Exception e) {
             return ResponseEntity.badRequest().body("Erro ao excluir veículo");
